Fixed Lem's code to work for VS community 2019
It had a problem defining an array with a number that wasn't a constant so switched them into a vector and filled the vectors with a for loop around this problem.  I also deleted a redundant folder that would've been used for an SFML environment with VS community 2019, but it didn't work out.