added hold option
when the hold bool is true, the drawn convex hull will stay on the screen after its finish iterating. if false it continues like it did before.
The user can continue to draw points on the graph after the program is done drawing and if they hit enter again it will clear the last drawn convex hull and draw another one.
You can set if hold is true or false by hitting P and O. P will set it to true and print "Hold is on" to console, and o will set it to false and print "Hold is off" to console.

also I deleted some commented out code that I forgot to get rid of when I was cleaning the push.