Added pause button
This creates a bool names pause. Essentially when pause is true it will stop the program from drawing anymore, pausing the program. The user can't make points when paused because the points wouldn't be seen until the program is un-paused and then they would all appear at the same time.
The user can toggle the pause bool by hitting the Spacebar.
The console also prints out if the program is paused or not paused when the user hits the space bar.