Bug Fix - Double Window
fixed a bug where two windows were being made with the one behind whitened out.  it was because technically a grahamscan object, which is responsible for holding the window. gets made in the header when it's declared. I was then also making one in the constructors so there were just two.