VisualConvexHull implemented into the Menu
I added visual convexHull to run when selected in the menu and I made the run menu that used to be included in its own main file into a runMenu function. with this implementation when the first option is selected in the menu it will run visual convex hull and after the visual convex hull window is closed it will run the menu function again.